refactor deliveralbe validator

+++ marlo-web/src/main/java/org/cgiar/ccafs/marlo/validation/projects/DeliverableValidator.java
@@ -401,7 +401,7 @@ public class DeliverableValidator extends BaseValidator {
     }
 
     if (deliverable.getPublication() != null && deliverable.getPublication().getId() != null
-      && deliverable.getPublication().getId() != -1) {
+      && deliverable.getPublication().getId().intValue() != -1) {
 
       DeliverablePublicationMetadata metadata = deliverable.getPublication();
